+++ src/apyfloatarray_wrapper.cc
@@ -521,6 +521,43 @@ void bind_float_array(nb::module_& m)
             )pbdoc"
         )
         .def_static(
+            "from_bits",
+            &APyFloatArray::from_bits,
+            nb::arg("bits"),
+            nb::arg("exp_bits"),
+            nb::arg("man_bits"),
+            nb::arg("bias") = nb::none(),
+            R"pbdoc(
+            Create an :class:`APyFloatArray` object from bit-representations.
+
+            Parameters
+            ----------
+            bits : int
+                The bit-representations.
+            exp_bits : int
+                Number of exponent bits in the created floating-point tensor
+            man_bits : int
+                Number of mantissa bits in the created floating-point tensor
+            bias : int, optional
+                Bias. If not provided, *bias* is ``2**exp_bits - 1``.
+
+            Examples
+            --------
+
+            >>> from apytypes import APyFloatArray
+
+            >>> a = APyFloatArray.from_bits(
+            ...     [[60, 61], [80, 82]],
+            ...     exp_bits=5,
+            ...     man_bits=2
+            ... )
+
+            Returns
+            -------
+            :class:`APyFloatArray`
+            )pbdoc"
+        )
+        .def_static(
             "zeros",
             &APyFloatArray::zeros,
             nb::arg("shape"),