+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   specifiers are given.
 - Added optional parameter for `bias` in array creation functions that were
   missing it.
+- Added `APyFloatArray.from_bits`.
 
 ### Fixed